finalizacion create team
